# backend/api/openapi.yaml
openapi: 3.1.3
info:
  title: The Special Standard API
  version: 0.1.0
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://king-prawn-app-n5vk6.ondigitalocean.app/api/v1
    description: Production server
...
paths:
...
  /sessions:
    get:
      summary: Get all sessions
      description: Retrieve all therapy sessions from the database
      tags: [Sessions]
      responses:
        "200":
          description: List of therapy sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: 404
                message: "Resource not found"

    post:
      summary: Create a new session
      description: Create a new therapy session
      tags: [Sessions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - start_datetime
                - end_datetime
                - therapist_id
              properties:
                start_datetime:
                  type: string
                  format: date-time
                  description: Start date and time of the session
                  example: "2025-09-13T17:19:00Z"
                end_datetime:
                  type: string
                  format: date-time
                  description: End date and time of the session
                  example: "2025-09-13T18:00:00Z"
                therapist_id:
                  type: string
                  format: uuid
                  description: UUID of the Therapist
                  example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                notes:
                  type: string
                  description: Optional notes about the session
                  example: "Daily Affirmation: My luck is so good that even four-leaf clovers ask for my autograph"
      responses:
        "201":
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          description: Bad Request (e.g., validation errors)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /sessions/{id}:
    get:
      summary: Get session by ID
      description: Retrieve a specific therapy session by its UUID
      tags: [Sessions]
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the Session to retrieve
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: Session found successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          description: Bad Request (invalid UUID format)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: 400
                message: "Invalid UUID format"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: 404
                message: "Session not found"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      summary: Update session
      description: Update an existing session (partial update)
      tags: [Sessions]
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the Session to patch
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSessionInput"
      responses:
        "200":
          description: Session updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          description: Bad request (e.g., validation errors)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Session Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete session
      description: Delete a session from the database
      tags: [Sessions]
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the Session to delete
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Session deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Session deleted successfully"
        "400":
          description: Bad Request Error (parsing invalid uuid, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /students:
    get:
      summary: Get all students
      description: Retrieve all students from the database
      tags: [Students]
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create a new student
      description: Create a new student record
      tags: [Students]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentInput"
      responses:
        "201":
          description: Student created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          description: Bad request (validation errors)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                invalid_json:
                  summary: Invalid JSON format
                  value:
                    code: 400
                    message: "Invalid JSON format"
                invalid_date:
                  summary: Invalid date format
                  value:
                    code: 400
                    message: "Invalid date format. Use YYYY-MM-DD"
                invalid_therapist_id:
                  summary: Invalid therapist ID
                  value:
                    code: 400
                    message: "Invalid therapist ID format"
        "422":
          description: Therapist not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: 422
                message: "Therapist not found"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /students/{id}:
    get:
      summary: Get student by ID
      description: Retrieve a specific student by their ID
      tags: [Students]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Student ID
          example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: Student found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: 400
                message: "Invalid UUID format"
        "404":
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: 404
                message: "Student not found"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      summary: Update student
      description: Update an existing student's information (partial update)
      tags: [Students]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Student ID
          example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentInput"
      responses:
        "200":
          description: Student updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          description: Bad request (validation errors)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                invalid_uuid:
                  summary: Invalid UUID format
                  value:
                    code: 400
                    message: "Invalid UUID format"
                invalid_json:
                  summary: Invalid JSON format
                  value:
                    code: 400
                    message: "Invalid JSON format"
                invalid_date:
                  summary: Invalid date format
                  value:
                    code: 400
                    message: "Invalid date format. Use YYYY-MM-DD"
        "404":
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: 404
                message: "Student not found"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete student
      description: Delete a student record
      tags: [Students]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Student ID
          example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "204":
          description: Student deleted successfully (no content)
        "400":
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: 400
                message: "Invalid UUID format"
        "404":
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: 404
                message: "Student not found"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /therapists:
    get:
      summary: Get all therapists
      description: Retrieve all therapists from the database
      tags: [Therapists]
      responses:
        "200":
          description: A list of Therapists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Therapist"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: 404
                message: "Resource not found"

    post:
      summary: Create a new therapist
      description: Create a new therapist for school districts
      tags: [Therapists]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTherapistInput"
      responses:
        "201":
          description: Therapist created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Therapist"
        "400":
          description: Bad request (e.g., validation errors)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /therapists/{id}:
    get:
      summary: Get therapist by ID
      description: Retrieve a specific therapist from the database
      tags: [Therapists]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Therapist ID
          example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: Therapist found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Therapist"
        "400":
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: 400
                message: "There was an error parsing the given id!"
        "404":
          description: Therapist not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: 404
                message: "Therapist not found"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      summary: Update therapist
      description: Update a specific field within a therapist object in the database
      tags: [Therapists]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Therapist ID
          example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTherapistInput"
      responses:
        "200":
          description: Therapist updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Therapist"
        "400":
          description: Bad request (validation errors)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: 400
                message: "There was an error parsing the given id!"
        "404":
          description: Therapist not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: 404
                message: "Therapist not found"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete therapist
      description: Get the specific therapist corresponding to the given ID and delete it
      tags: [Therapists]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Therapist ID
          example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: Therapist deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Therapist deleted successfully"
        "400":
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: 400
                message: "Error querying database for given ID"
        "404":
          description: Therapist not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: 404
                message: "Therapist not found"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /resources:
    get:
      summary: Get all resources
      description: Retrieve all resources from the database
      tags: [Resources]
      parameters:
        - in: query
          name: theme_id
          schema:
            type: string
            format: uuid
          required: false
          description: Filter resources by theme ID
        - in: query
          name: grade_level
          schema:
            type: string
          required: false
          description: Filter resources by grade level
        - in: query
          name: date
          schema:
            type: string
            format: date
          required: false
          description: Filter resources by date
        - in: query
          name: type
          schema:
            type: string
          required: false
          description: Filter resources by type
        - in: query
          name: title
          schema:
            type: string
          required: false
          description: Filter resources by title
        - in: query
          name: category
          schema:
            type: string
          required: false
          description: Filter resources by category
        - in: query
          name: content
          schema:
            type: string
          required: false
          description: Filter resources by content/URL
      responses:
        "200":
          description: List of resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resource"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: 404
                message: "Resource not found"
    post:
      summary: Create a new resource
      description: Create a new resource in the database
      tags: [Resources]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResourceBody"
      responses:
        "201":
          description: Resource created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "400":
          description: Bad request (e.g., validation errors)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found (e.g. theme_id does not exist)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /resources/:id:
    get:
      summary: Get resource by ID
      description: Retrieve a specific resource by its ID
      tags: [Resources]
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: ID of the resource to retrieve
      responses:
        "200":
          description: Resource found successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "400":
          description: Bad request (e.g., invalid UUID format)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      summary: Update a resource
      description: Update an existing resource in the database
      tags: [Resources]
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: ID of the resource to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResourceBody"
      responses:
        "200":
          description: Resource updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "400":
          description: Bad request (e.g., validation errors)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a resource
      description: Delete an existing resource from the database
      tags: [Resources]
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: ID of the resource to delete
      responses:
        "204":
          description: Resource deleted successfully (no content)
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /session-resource:
    post:
      summary: Create a session-resource link
      description: Create a new association between each session and a utilized resource
      tags: [SessionResource]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModifySessionResource"
      responses:
        "200":
          description: Session-resource link created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResource"
              example:
                id: "123e4567-e89b-12d3-a456-426614174000"
                session_id: "123e4567-e89b-12d3-a456-426614174001"
                resource_id: "123e4567-e89b-12d3-a456-426614174002"
                created_at: "2025-09-13T17:19:00Z"
                updated_at: "2025-09-13T17:19:00Z"
        "400":
          description: Bad request (e.g., validation errors)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
                example:
                  code: 400
                  message: "Reference not found"
        "404":
          description: Session or Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
                example:
                  code: 404
                  message: "Session or Resource not found"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
                example:
                  session_id: "123e4567-e89b-12d3-a456-426614174001"
                  resource_id: "123e4567-e89b-12d3-a456-426614174002"
    delete:
      summary: Delete a session-resource link
      description: Delete an existing association between a session and a resource
      tags: [SessionResource]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModifySessionResource"
              example:
                session_id: "123e4567-e89b-12d3-a456-426614174001"
                resource_id: "123e4567-e89b-12d3-a456-426614174002"
        responses:
          "204":
            description: Session-resource link deleted successfully (no content)
          "404":
            description: Session-resource link not found
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Error"
                  example:
                    code: 404
                    message: "Session or Resource not found"
          "400":
            description: Bad request (e.g., validation errors)
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Error"
                  example:
                    code: 400
                    message: "Reference not found"
          "500":
            description: Internal server error
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 500
        message:
          anyOf:
            - type: string
            - type: object
          description: Error message or validation errors
          example: "Internal server error"
...
    CreateTherapistInput:
      type: object
      required:
        - first_name
        - last_name
        - email
      properties:
        first_name:
          type: string
          description: First name of the therapist
          example: "Kevin"
        last_name:
          type: string
          description: Last name of the therapist
          example: "Matilda"
        email:
          type: string
          format: email
          description: Email of the therapist
          example: "email123@example.com"
...
    Session:
      type: object
      required:
        - id
        - start_datetime
        - end_datetime
        - therapist_id
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the session
          example: "123e4567-e89b-12d3-a456-426614174000"
        start_datetime:
          type: string
          format: date-time
          description: Start date and time of the session
          example: "2024-01-15T14:30:00Z"
        end_datetime:
          type: string
          format: date-time
          description: End date and time of the session
          example: "2024-01-15T15:30:00Z"
        therapist_id:
          type: string
          format: uuid
          description: ID of the therapist conducting the session
          example: "987fcdeb-51a2-43d1-9f4c-123456789abc"
        notes:
          type: string
          nullable: true
          description: Session notes (optional)
          example: "Patient showed improvement in anxiety management techniques"
        created_at:
          type: string
          format: date-time
          nullable: true
          description: When the session record was created
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          nullable: true
          description: When the session record was last updated
          example: "2024-01-15T10:35:00Z"
...